// packages/front/src/shared/lib/parsers/types.ts
import { EPlaylistDataType } from "shared/types";

export type THtmlPlaylistInfo = {
    id: string;
    subtitle: string;
};

export type THtmlPlaylistBlock = {
    blockId: string;
    title: string;
    dataType: EPlaylistDataType;
    showAllLink: string;
    coverPlaylists: THtmlPlaylistInfo[];
};